+++ aws-android-sdk-mobile-client/src/main/java/com/amazonaws/mobile/client/AWSMobileClient.java
@@ -3790,111 +3790,6 @@ public final class AWSMobileClient implements AWSCredentialsProvider {
     }
 }
 
-interface KeyValueStore {
-    Map<String, String> get(final String... keys);
-    String get(final String key);
-    void set(final Map<String, String> attributes);
-    void set(final String key, final String value);
-    void clear();
-}
-
-final class DummyStore implements KeyValueStore {
-    @Override
-    public Map<String, String> get(String... keys) {
-        throwNotInitializedException();
-        return null;
-    }
-
-    @Override
-    public String get(String key) {
-        throwNotInitializedException();
-        return null;
-    }
-
-    @Override
-    public void set(Map<String, String> attributes) {
-        throwNotInitializedException();
-    }
-
-    @Override
-    public void set(String key, String value) {
-        throwNotInitializedException();
-
-    }
-
-    @Override
-    public void clear() {
-        throwNotInitializedException();
-    }
-
-    private void throwNotInitializedException() {
-        throw new RuntimeException("AWSMobileClient has not been initialized yet.");
-    }
-}
-
-final class AWSMobileClientStore implements KeyValueStore {
-    AWSKeyValueStore mAWSKeyValueStore;
-
-    private ReadWriteLock mReadWriteLock = new ReentrantReadWriteLock();
-
-    AWSMobileClientStore(AWSMobileClient client) {
-        mAWSKeyValueStore = new AWSKeyValueStore(client.mContext,
-                AWSMobileClient.SHARED_PREFERENCES_KEY,
-                client.mIsPersistenceEnabled);
-    }
-
-    @Override
-    public Map<String, String> get(final String... keys) {
-        try {
-            mReadWriteLock.readLock().lock();
-            HashMap<String, String> attributes = new HashMap<String, String>();
-            for (String key : keys) {
-                attributes.put(key, mAWSKeyValueStore.get(key));
-            }
-            return attributes;
-        } finally {
-            mReadWriteLock.readLock().unlock();
-        }
-    }
-
-    @Override
-    public String get(final String key) {
-        try {
-            mReadWriteLock.readLock().lock();
-            return mAWSKeyValueStore.get(key);
-        } finally {
-            mReadWriteLock.readLock().unlock();
-        }
-    }
-
-    @Override
-    public void set(final Map<String, String> attributes) {
-        try {
-            mReadWriteLock.writeLock().lock();
-            for (String key : attributes.keySet()) {
-                mAWSKeyValueStore.put(key, attributes.get(key));
-            }
-        } finally {
-            mReadWriteLock.writeLock().unlock();
-        }
-    }
-
-    @Override
-    public void set(final String key, final String value) {
-        try {
-            mReadWriteLock.writeLock().lock();
-            mAWSKeyValueStore.put(key, value);
-        } finally {
-            mReadWriteLock.writeLock().unlock();
-        }
-    }
-
-    @Override
-    public void clear() {
-        mAWSKeyValueStore.clear();
-    }
-}
-
 /**
  * A duplicate class of AWSEnhancedCognitoIdentityProvider that provides the ability to
  * branch into developer authenticated identities.
