+++ aws-android-sdk-mobile-client/src/main/java/com/amazonaws/mobile/client/KeyValueStore.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2017-2021 Amazon.com, Inc. or its affiliates.
+ * All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.amazonaws.mobile.client;
+
+import java.util.Map;
+
+/**
+ * Allows key-value pairs to be stored and retrieved.
+ */
+interface KeyValueStore {
+    Map<String, String> get(final String... keys);
+    String get(final String key);
+    void set(final Map<String, String> attributes);
+    void set(final String key, final String value);
+    void clear();
+}
